Extract some code out of the main Middleware class into a new BaseMiddleware class.
