+++ lib/shrimp/middleware.rb
@@ -1,58 +1,48 @@
+require 'shrimp/base_middleware'
+
 module Shrimp
-  class Middleware
+  class Middleware < BaseMiddleware
     def initialize(app, options = { }, conditions = { })
-      @app                        = app
-      @options                    = options
-      @conditions                 = conditions
+      super
       @options[:polling_interval] ||= 1
       @options[:polling_offset]   ||= 1
       @options[:cache_ttl]        ||= 1
       @options[:request_timeout]  ||= @options[:polling_interval] * 10
     end
 
-    def call(env)
-      @request = Rack::Request.new(env)
-      if render_as_pdf? #&& headers['Content-Type'] =~ /text\/html|application\/xhtml\+xml/
-        if already_rendered? && (up_to_date?(@options[:cache_ttl]) || @options[:cache_ttl] == 0)
-          if File.size(render_to) == 0
-            File.delete(render_to)
-            remove_rendering_flag
-            return error_response
-          end
-          return ready_response if env['HTTP_X_REQUESTED_WITH']
-          file = File.open(render_to, "rb")
-          body = file.read
-          file.close
-          File.delete(render_to) if @options[:cache_ttl] == 0
+    def render_as_pdf(env)
+      if already_rendered? && (up_to_date?(@options[:cache_ttl]) || @options[:cache_ttl] == 0)
+        if File.size(render_to) == 0
+          File.delete(render_to)
           remove_rendering_flag
-          response                  = [body]
-          headers                   = { }
-          headers["Content-Length"] = (body.respond_to?(:bytesize) ? body.bytesize : body.size).to_s
-          headers["Content-Type"]   = "application/pdf"
-          [200, headers, response]
-        else
-          if rendering_in_progress?
-            if rendering_timed_out?
-              remove_rendering_flag
-              error_response
-            else
-              reload_response(@options[:polling_interval])
-            end
-          else
-            File.delete(render_to) if already_rendered?
-            set_rendering_flag
-            fire_phantom
-            reload_response(@options[:polling_offset])
-          end
+          return error_response
         end
+        return ready_response if env['HTTP_X_REQUESTED_WITH']
+        file = File.open(render_to, "rb")
+        body = file.read
+        file.close
+        File.delete(render_to) if @options[:cache_ttl] == 0
+        remove_rendering_flag
+        response                  = [body]
+        headers                   = { }
+        headers["Content-Length"] = (body.respond_to?(:bytesize) ? body.bytesize : body.size).to_s
+        headers["Content-Type"]   = "application/pdf"
+        [200, headers, response]
       else
-        @app.call(env)
+        if rendering_in_progress?
+          if rendering_timed_out?
+            remove_rendering_flag
+            error_response
+          else
+            reload_response(@options[:polling_interval])
+          end
+        else
+          File.delete(render_to) if already_rendered?
+          set_rendering_flag
+          fire_phantom
+          reload_response(@options[:polling_offset])
+        end
       end
-    end
-
-    # The URL for the HTML-formatted web page that we are converting into a PDF.
-    def html_url
-      @request.url.sub(%r<\.pdf(\?|$)>, '\1')
     end
 
     private
@@ -61,12 +51,6 @@ module Shrimp
     def fire_phantom
       puts %(#{self.class}: Converting web page at #{(html_url).inspect} into a PDF ...) if Shrimp.configuration.default_options[:debug]
       Process::detach fork { Phantom.new(html_url, @options, @request.cookies).to_pdf(render_to) }
-    end
-
-    def render_to
-      file_name = Digest::MD5.hexdigest(@request.path) + ".pdf"
-      file_path = @options[:out_path]
-      "#{file_path}/#{file_name}"
     end
 
     def already_rendered?
@@ -97,36 +81,6 @@ module Shrimp
       @request.session["phantom-rendering"][render_to]
     end
 
-    def render_as_pdf?
-      request_path_is_pdf = !!@request.path.match(%r{\.pdf$})
-
-      if request_path_is_pdf && @conditions[:only]
-        rules = [@conditions[:only]].flatten
-        rules.any? do |pattern|
-          if pattern.is_a?(Regexp)
-            @request.path =~ pattern
-          else
-            @request.path[0, pattern.length] == pattern
-          end
-        end
-      elsif request_path_is_pdf && @conditions[:except]
-        rules = [@conditions[:except]].flatten
-        rules.map do |pattern|
-          if pattern.is_a?(Regexp)
-            return false if @request.path =~ pattern
-          else
-            return false if @request.path[0, pattern.length] == pattern
-          end
-        end
-        return true
-      else
-        request_path_is_pdf
-      end
-    end
-
-    def concat(accepts, type)
-      (accepts || '').split(',').unshift(type).compact.join(',')
-    end
 
     def reload_response(interval=1)
       body = <<-HTML.gsub(/[ \n]+/, ' ').strip
